fix: typecheck – ticker route and Discord state route exports
- Ticker API: use requested.length (not .size); safer timestamp handling for posts fallback
- Discord activity: move generateSignedState/verifySignedState to state-utils.ts so route only exports GET (fixes Next.js route type constraint)
